[fixup][win] Fix QtWebEngine build on Windows
Reviewed-on: https://codereview.qt-project.org/c/qt/qtwebengine-chromium/+/698814

+++ chromium/v8/src/utils/utils.h
@@ -258,7 +258,7 @@ inline T RoundingAverageUnsigned(T a, T b) {
 // Defines padding field with given names which aligns next field by Alignment
 // bytes, assuming field definition macro V(CamelName, Size, hacker_name).
 #define PADDING_FIELD(Alignment, V, Name, hacker_name) \
-  V(Name, (RoundUp<Alignment>(k##Name##Offset) - k##Name##Offset), hacker_name)
+  V(Name, (RoundUp<Alignment>(static_cast<int>(k##Name##Offset)) - k##Name##Offset), hacker_name)
 
 // Size of the field defined by DEFINE_FIELD_OFFSET_CONSTANTS
 #define FIELD_SIZE(Name) (Name##End + 1 - Name)
